Extract abbreviation logic into utils
